% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe2netcdf.R
\name{dataframe2netcdf}
\alias{dataframe2netcdf}
\title{Writing a data frame to a netcdf file.}
\usage{
dataframe2netcdf(data_frame, netcdf_file,
  dim_names = intersect(names(data_frame), c("lon", "lat", "longitude",
  "latitude", "time")), dim_units = NULL,
  var_names = names(data_frame)[!names(data_frame) \%in\% dim_names],
  var_units = rep("", length(var_names)), var_longnames = var_names,
  overwrite_existing = FALSE, fill_dimensions = FALSE,
  max_resolution_testing = 5, verbose = FALSE)
}
\arguments{
\item{data_frame}{Data frame that includes dimension names and at least one variable.}

\item{netcdf_file}{File path for netcdf file to write to.}

\item{dim_names}{Names of dimensions in data frame.}

\item{dim_units}{Units for all dimensions.}

\item{var_names}{Names of variables in data frame.}

\item{var_units}{Units for all variables.}

\item{var_longnames}{Longnames for each variable (default is variable name).}

\item{overwrite_existing}{If file exists already, should it be overwritten?}

\item{fill_dimensions}{If TRUE, dimensions will be extended to cover the whole world.}

\item{max_resolution_testing}{Maximum number of resolutions to be tested, for dimension filling.}

\item{verbose}{Include print statements.}
}
\value{
Returns file name of netcdf file, if saved successfully.
}
\description{
Writing a data frame to a netcdf file.
}
