% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcdf2dataframe.R
\name{netcdf2dataframe}
\alias{netcdf2dataframe}
\title{Reading a netcdf file into a data frame.}
\usage{
netcdf2dataframe(netcdf_file, variables = "all", remove_NA = FALSE,
  time_format = "\%Y-\%m-\%d \%H:\%M:\%S", start_idx = NULL,
  end_idx = NULL, count = NULL, grid_cells = NULL, years = NULL,
  return_time_columns = FALSE, verbose = FALSE)
}
\arguments{
\item{netcdf_file}{Path to a netcdf file.}

\item{variables}{Vector of variable names to read in.}

\item{remove_NA}{Specify if empty observations should be removed.}

\item{time_format}{Format vector. Specify how time variable should be returned.}

\item{start_idx}{Specify the start index of the read in.}

\item{end_idx, count}{Specify either the end index (end_idx) or the number of steps to read in (count).}

\item{grid_cells}{Data frame with lat/lon columns that specify the coordinates to read in.}

\item{years}{Specify which years should be read in.}

\item{verbose}{Include print statements.}
}
\value{
A data frame with the dimensions and variables stored in the netcdf file, as specified.
}
\description{
Reading a netcdf file into a data frame.
}
\examples{
# Download a file from the HadCRU dataset:
url = "https://crudata.uea.ac.uk/cru/data/temperature/HadSST.3.1.1.0.median.nc"
netcdf_file = "HadSST.3.1.1.0.median.nc"
download.file(url, netcdf_file)

df = netcdf2dataframe(netcdf_file,
                      variables = "sst",
                      time_format = "\%Y-\%m-\%d")
df = netcdf2dataframe(netcdf_file,
                      variables = "sst",
                      time_format = "\%Y-\%m-\%d",
                      remove_NA = TRUE)
df = netcdf2dataframe(netcdf_file,
                      variables = "sst",
                      time_format = "\%Y-\%m-\%d",
                      years = 2000)
df = netcdf2dataframe(netcdf_file,
                      variables = "sst",
                      time_format = "\%Y-\%m-\%d",
                      years = 2000:2010,
                      grid_cells = data.frame(
                      lon = -17.5, lat = -72.5))
}
